requirements.txt
fastapi
uvicorn
httpx
google-api-python-client
google-auth
python-dotenv
profanity-check
